% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MANDALA-package.R
\docType{package}
\name{MANDALA-package}
\alias{MANDALA-package}
\title{MANDALA: Multiscale Aggregation and Neighborhood Data Analysis for Local Architecture in Spatial and Single Cell RNA Sequencing Study}
\description{
MANDALA provides intuitive solutions to visualizing differences among cells in single-cell and spatial RNA-seq datasets. Cells are aggregated into bubbles or hexagonal bins before rendering, so identity, local density, gene expression and metadata remain legible at atlas scale. Interactive viewers are written as self-contained HTML and allow exports into vector or high res PNG. Runs locally in R on Seurat v3, v4 and v5 objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/merlab/MANDALA}
  \item Report bugs at \url{https://github.com/merlab/MANDALA/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Ho \email{nicolas.ho@uottawa.ca} (\href{https://orcid.org/0000-0002-3192-3049}{ORCID})

Authors:
\itemize{
  \item Arvind Mer \email{amer@uottawa.ca} (\href{https://orcid.org/0000-0001-8510-7963}{ORCID}) [funder]
}

}
\keyword{internal}
